+++ Gini_Splitter.cpp
@@ -2,7 +2,7 @@
  * @ Author: Adam Myers
  * @ Create Time: 2023-04-11 11:29:14
  * @ Modified by: Adam Myers
- * @ Modified time: 2023-04-11 15:17:36
+ * @ Modified time: 2023-04-13 17:33:58
  * @ Description: Implementation of the Gini_Splitter derived class.
  */
 
@@ -10,7 +10,7 @@
 #include "Gini_Splitter.h"
 
 Gini_Splitter::~Gini_Splitter() {};
-double Gini_Splitter::calculate_node_score(const std::vector<std::pair<std::vector<float>, float>>& data, const std::vector<size_t> indexes) const
+double Gini_Splitter::calculate_node_score(const std::vector<std::pair<std::vector<float>, int>>& data, const std::vector<size_t> indexes) const
     {
         std::unordered_map<int, std::size_t> label_counts_map = label_counts(data, indexes);
         double n_data_points = indexes.size();
