+++ Gini_Splitter.h
@@ -2,7 +2,7 @@
  * @ Author: Adam Myers
  * @ Create Time: 2023-04-11 11:29:08
  * @ Modified by: Adam Myers
- * @ Modified time: 2023-04-11 15:19:32
+ * @ Modified time: 2023-04-13 17:34:09
  * @ Description: Header file for the Gini_Splitter derived class.
  * 
  * This class inherets all the utility functions of the Splitter base class 
@@ -20,7 +20,7 @@ class Gini_Splitter : public Splitter
 {
 public:
     virtual ~Gini_Splitter();
-    double calculate_node_score(const std::vector<std::pair<std::vector<float>, float>>& data, const std::vector<size_t> indexes) const;
+    double calculate_node_score(const std::vector<std::pair<std::vector<float>, int>>& data, const std::vector<size_t> indexes) const;
 };
 
 #endif
